requirements.txt
opencv-python==4.1.0.25
tesserocr==2.4.0
numpy==1.16.3
Pillow==6.0.0
imutils==0.5.2